Rename display.py to print_updates.py
